% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_mean_difference.R
\name{compute_sub_smd}
\alias{compute_sub_smd}
\alias{SMD}
\title{Compute standardized mean differences for subclassification result}
\usage{
compute_sub_smd(mi_obj, sd = "pooled")
}
\arguments{
\item{mi_obj}{A matchit object derived from MatchIt pacakge}

\item{sd}{The standard deviation used as the denominator in the formula}
}
\value{
Return a scalar (the overall SMD)
}
\description{
This function accepts a MatchIt object (i.e., the result of matchit function)
, and calculates the overall standardized mean difference after
subclassification. Note only subclassification result is applicable to this
function. For other matching results except for exact matching,
use compute_smd() instead. In addition, SMD can be calculated on the basis of
the standard deviation of original treatment group, which is the formula
used in matchit function, or on the basis of the simple pooled standard
deviation of original treatment and control group. The default is sd =
"pooled", but it can be switched to "treatment".
}
\examples{
take lalonde data as an example
run \code{\link[MatchIt]{matchit}} to obtain the matching result
> \code{m_out <- matchit(treat ~ re74 + re75 + age + educ + hisp + black,
data = lalonde, method = "subclass", subclass = 7)}
> \code{compute_sub_smd(m_out, sd = "treatment")}
}
\seealso{
compute_smd()
}
\keyword{SMD}
\keyword{subclassification}
