% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula to obtain grouping variable and covariate vector}
\usage{
parse_formula(mi_obj = NULL)
}
\arguments{
\item{mi_obj}{A matchit object derived from MatchIt pacakge}
}
\value{
Return a list including the grouping variable and covariate vector
}
\description{
This function parses the formula used in matchit() to obtain the grouping
variable and the covariate vector.
}
\examples{
m_out <- MatchIt::matchit(treat ~ re74 + re75 + age + educ + hispan +
black, data = MatchIt::lalonde, method = "nearest")
parse_formula(m_out)
}
\keyword{covariate}
\keyword{grouping}
